Moved functions related to creating pin from machine to customer
