+++ include/customer.h
@@ -8,4 +8,6 @@
 int get_customer_info();   
 int generate_account();
 int write_user_data(std::string first, std::string last, int acct_num, int pin); 
+int create_pin();
+int verify_pin(int unverified_pin);
 #endif
