+++ include/machine.h
@@ -4,7 +4,5 @@
 #include <iostream>
 
 void welcome_menu();
-int create_pin();
-int verify_pin(int unverified_pin);
     
 #endif
